requirements.txt
# Requirements. These can can be installed with:
#   pip install -r requirements.txt

# You will need a C compiler and proper setup for Cython
# https://wiki.python.org/moin/WindowsCompilers


    numpy>=1.20.1
    voluptuous>=0.12.0
    pyaudio>=0.2.11
    sacn==1.5.0
    aiohttp~=3.7.4.post0
    multidict>=5.0.0
    aiohttp_jinja2>=1.1.0
    requests>=2.24.0
    pyyaml>=5.3.1
    aubio>=0.4.9
    zeroconf<=0.28.8
    cython==0.29.21
    pyupdater>=4.0.0
    sentry-sdk==1.0.0
    certifi>=2019.3.9
    pyserial>=3.5
    python-rtmidi>=1.4.9
